+++ src/main/java/com/tristankechlo/toolleveling/commands/ToolLevelingCommand.java
@@ -33,17 +33,14 @@ public final class ToolLevelingCommand {
     private static int configReload(CommandContext<ServerCommandSource> context) {
         ServerCommandSource source = context.getSource();
         ConfigManager.reloadAllConfigs(source.getServer());
-        source.sendFeedback(Text.translatable("commands.toolleveling.config.reload"), true);
+        ResponseHelper.sendMessageConfigReload(source);
         return 1;
     }
 
     private static int configShow(CommandContext<ServerCommandSource> context) {
         ServerCommandSource source = context.getSource();
         final ConfigIdentifier config = context.getArgument("identifier", ConfigIdentifier.class);
-        String name = config.getFileName();
-        String path = ConfigManager.getConfigPath(config);
-        MutableText mutableText = (Text.literal(name)).formatted(Formatting.UNDERLINE).styled(style -> style.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_FILE, path)));
-        source.sendFeedback(Text.translatable("commands.toolleveling.config.path", mutableText), true);
+        ResponseHelper.sendMessageConfigShow(source, config);
         return 1;
     }
 
@@ -51,7 +48,7 @@ public final class ToolLevelingCommand {
         ServerCommandSource source = context.getSource();
         final ConfigIdentifier config = context.getArgument("identifier", ConfigIdentifier.class);
         ConfigManager.resetOneConfig(source.getServer(), config);
-        source.sendFeedback(Text.translatable("commands.toolleveling.config.reset", config.withModID()), true);
+        ResponseHelper.sendMessageConfigReset(source, config);
         return 1;
     }