+++ Changelog.md
@@ -17,6 +17,7 @@
 - if both `global_minimum_enchantment_level` and `minimum_enchantment_levels` are set, the higher of the two will be used
 - update in-game mod icon to the newer blockmodel
 - add new command `/toolleveling config info <config>` to display links to the config-explanation
+- create consistent formatting for command responses
 
 ### Version 1.19 - 1.3.3
  - rename config-identifier from `itemvalues` to `item_values`